Rewrite comp.c
Because the comparison code computes the number of common fingerprints
wrongly, I eventually decided to rewrite the comp program.
In the original, if there is a common hash in both sets, it sums up the number
of occurences of this hash of both sets.
As a result, e.g. the hashes like 'a a a a b' and 'a b b b b' would
score 100% similarity.